tidied up authorisation so that Members can access community related data
