+++ src/domain/user-group/user-group.resolver.fields.ts
@@ -7,6 +7,7 @@ import { RestrictedGroupNames, UserGroup } from './user-group.entity';
 import { UserGroupService } from './user-group.service';
 import { UserGroupParent } from './user-group-parent.dto';
 import { Profiling } from '@utils/logging/logging.profiling.decorator';
+import { User } from '@domain/user/user.entity';
 
 @Resolver(() => UserGroup)
 export class UserGroupResolverFields {
@@ -21,5 +22,16 @@ export class UserGroupResolverFields {
   @Profiling.api
   async parent(@Parent() userGroup: UserGroup) {
     return await this.userGroupService.getParent(userGroup);
+  }
+
+  @Roles(RestrictedGroupNames.Members)
+  @UseGuards(GqlAuthGuard)
+  @ResolveField('members', () => User)
+  @Profiling.api
+  async members(@Parent() group: UserGroup): Promise<User[]> {
+    if (!group || !group.membersPopulationEnabled) return [];
+
+    const members = await this.userGroupService.getMembers(group.id);
+    return (members || []) as User[];
   }
 }
