+++ src/domain/user-group/user-group.resolver.ts
@@ -1,17 +1,10 @@
 import { UseGuards } from '@nestjs/common';
-import {
-  Args,
-  Mutation,
-  Parent,
-  ResolveField,
-  Resolver,
-} from '@nestjs/graphql';
+import { Args, Mutation, Resolver } from '@nestjs/graphql';
 import { Roles } from '@utils/decorators/roles.decorator';
 import { GqlAuthGuard } from '@utils/authentication/graphql.guard';
 import { RestrictedGroupNames, UserGroup } from './user-group.entity';
 import { IUserGroup } from './user-group.interface';
 import { UserGroupService } from './user-group.service';
-import { User } from '@domain/user/user.entity';
 import { Profiling } from '@utils/logging/logging.profiling.decorator';
 
 @Resolver(() => UserGroup)
@@ -88,14 +81,5 @@ export class UserGroupResolver {
   ): Promise<IUserGroup> {
     const group = await this.groupService.removeFocalPoint(groupID);
     return group;
-  }
-
-  @ResolveField('members', () => User)
-  @Profiling.api
-  async members(@Parent() group: UserGroup): Promise<User[]> {
-    if (!group || !group.membersPopulationEnabled) return [];
-
-    const members = await this.groupService.getMembers(group.id);
-    return (members || []) as User[];
   }
 }
